fixed handlers
fixed the bottleStatus (on/off) and water level pins and handlers on
the device side. right now it doesn’t display that info on the phone
side though. Some of the handlers needed on the phone side are there
but are not properly implemented. IMPORTANT TO REMEMBER: only the
on/off deviceStatus should be a two-way street, waterLevel should only
be changed from the device side and should not be changed ever from the
phone.